Add SkiaMetalSurfaceFactory
