+++ package/ios/RNSkia-iOS/RNSkMetalCanvasProvider.h
@@ -13,16 +13,6 @@
 
 #pragma clang diagnostic pop
 
-using SkiaMetalContext = struct SkiaMetalContext {
-  id<MTLCommandQueue> commandQueue = nullptr;
-  sk_sp<GrDirectContext> skContext = nullptr;
-};
-
-class ThreadContextHolder {
-public:
-  static thread_local SkiaMetalContext ThreadMetalContext;
-};
-
 class RNSkMetalCanvasProvider : public RNSkia::RNSkCanvasProvider {
 public:
   RNSkMetalCanvasProvider(std::function<void()> requestRedraw,
@@ -38,17 +28,7 @@ public:
   void setSize(int width, int height);
 
   CALayer *getLayer();
-
-  static sk_sp<SkSurface> MakeOffscreenMetalSurface(int width, int height);
-
 private:
-  /**
-   * To be able to use static contexts (and avoid reloading the skia context for
-   * each new view, we track the Skia drawing context per thread.
-   * @return The drawing context for the current thread
-   */
-  static SkiaMetalContext &createSkiaDirectContextIfNecessary();
-  static id<MTLDevice> device;
   std::shared_ptr<RNSkia::RNSkPlatformContext> _context;
   float _width = -1;
   float _height = -1;
